Component Work Pt 1
Most of the component side programming is complete. We can now send a message to the NPC, and via breakpoints we do see that the info is being processed by the plugin correctly. We just need to debug why we are not getting an output, then move on to the final step of pushing that out to the UI.